+++ Sources/KeyboardKit/Layout/Providers/SystemKeyboardLayoutProvider+Utils.swift
@@ -43,34 +43,6 @@ public extension SystemKeyboardLayoutProvider {
     func isAlphabeticWithInputCount(_ context: KeyboardContext, _ rows: [Int]) -> Bool {
         context.isAlphabetic && hasAlphabeticInputCount(rows)
     }
-
-    /**
-     Get the leading margin action for a certain action row.
-     */
-    func leadingMarginAction(for actions: KeyboardActions) -> KeyboardAction {
-        marginAction(for: actions.first { $0.isInputAction })
-    }
-
-    /**
-     Get the trailing margin action for a certain action row.
-     */
-    func trailingMarginAction(for actions: KeyboardActions) -> KeyboardAction {
-        marginAction(for: actions.last { $0.isInputAction })
-    }
-
-    /**
-     Get a margin action for a certain action, if any.
-
-     This function returns `characterMargin` for `character`
-     and `none` for all other action types.
-     */
-    func marginAction(for action: KeyboardAction?) -> KeyboardAction {
-        guard let action = action else { return .none }
-        switch action {
-        case .character(let char): return .characterMargin(char)
-        default: return .none
-        }
-    }
 }