+++ Sources/KeyboardKit/_Deprecated/KeyboardAction+Deprecated.swift
@@ -8,3 +8,15 @@ public extension KeyboardAction {
     @available(*, deprecated, renamed: "isUppercasedShiftAction")
     var isUppercaseShift: Bool { isUppercasedShiftAction }
 }
+
+/**
+ This extension makes this enum implement ``KeyboardRowItem``.
+ */
+extension KeyboardAction {
+
+    /**
+     The row-specific ID to use when the action is presented
+     in a keyboard row.
+     */
+    public var rowId: KeyboardAction { self }
+}